requirements.txt
Django==1.11.29
django-cors-headers==3.0.2
django-restframework==0.0.1
djangorestframework==3.9.4
python-jose==3.2.0
